requirements.txt
django==4.1.3
djangorestframework==3.14.0
argon2-cffi==21.3
gunicorn==20.1.0
drf-yasg==1.21.4
faker==15.3.4
psycopg2==2.9.5
django-trench==0.3.1
django-phonenumber-field[phonenumbers]
